# Conflicts:
#	resources/views/laka-user-management/add_contact_update.blade.php

+++ resources/views/components/laka-user-management/btn-accept.blade.php
@@ -23,13 +23,17 @@
     $nameBtn = $arrNameBtn[$data['request_approval_status']];
     $colorBtn = $arrColorBtn[$data['request_approval_status']];
     $iconBtn = $arrIcon[$data['request_approval_status']];
+
 @endphp
 {!! bt_link_to_route($urlAction, __("common.$nameBtn"), $colorBtn, ['id'=>$data['id']], ['class' => 'btn-sm', 'icon' => $iconBtn,
-'onclick'=>"return window.confirm('".__('common.are_you_sure',['message'=>"{$nameBtn}"])."')"]) !!}
+"data-confirmation-msg"=>__('common.are_you_sure',['message'=>"{$nameBtn}"])],'edit',$sectionCode) !!}
 
-<a name="destroy" class="btn btn-sm btn-danger data-remote"
-   href="{{ route("{$sectionCode}.delete-token",['id'=>$data['id']])}}" title="Delete" data-loading="Loading"
-   data-trigger-confirm="1" data-confirmation-msg="Are you sure you want to delete?" data-method="DELETE"
-   data-pjax-target="#approvalApiToken-grid">
-    <i class="fa fa fa-trash"></i>
-</a>
+@if(user_can("edit_$sectionCode"))
+    <a name="destroy" class="btn btn-sm btn-danger data-remote"
+       href="{{ route("{$sectionCode}.delete-token",['id'=>$data['id']])}}" title="Delete" data-loading="Loading"
+       data-trigger-confirm="1" data-confirmation-msg="Are you sure you want to delete?" data-method="DELETE"
+       data-pjax-target="#approvalApiToken-grid">
+        <i class="fa fa fa-trash"></i>
+    </a>
+@endif
+
